# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brain Bridge API Documentation",
        "description": "API documentation for Brain Bridge application",
        "contact": {
            "email": "gmzesan7767@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user information",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user information",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "This is my bio"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St, City, Country"
                                    },
                                    "profile_picture": {
                                        "type": "string",
                                        "example": "http://example.com/profile.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/video-lessons": {
            "get": {
                "tags": [
                    "Video Lessons"
                ],
                "summary": "Get list of all video lessons",
                "description": "Returns all video lessons from the database",
                "operationId": "getVideoLessonsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Video Lessons"
                ],
                "summary": "Create a new video lesson",
                "description": "Creates a new video lesson",
                "operationId": "storeVideoLesson",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "teacher_id",
                                    "subject_id",
                                    "title",
                                    "video_url"
                                ],
                                "properties": {
                                    "teacher_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subject_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Introduction to Algebra"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A basic introduction to algebraic concepts."
                                    },
                                    "old_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 49.99
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 29.99
                                    },
                                    "duration_hours": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "video_url": {
                                        "type": "string",
                                        "example": "https://example.com/video.mp4"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video lesson created successfully"
                    }
                }
            }
        },
        "/api/video-lessons/{id}": {
            "get": {
                "tags": [
                    "Video Lessons"
                ],
                "summary": "Get video lesson information",
                "description": "Returns video lesson data",
                "operationId": "getVideoLessonById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video Lesson id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Video Lesson not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Video Lessons"
                ],
                "summary": "Update an existing video lesson",
                "description": "Updates a video lesson by ID",
                "operationId": "updateVideoLesson",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video Lesson ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "old_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "duration_hours": {
                                        "type": "integer"
                                    },
                                    "video_url": {
                                        "type": "string"
                                    },
                                    "is_published": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Video lesson updated successfully"
                    },
                    "404": {
                        "description": "Video Lesson not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Video Lessons"
                ],
                "summary": "Delete a video lesson",
                "description": "Deletes a video lesson by ID",
                "operationId": "deleteVideoLesson",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video Lesson ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video lesson deleted successfully"
                    },
                    "404": {
                        "description": "Video Lesson not found"
                    }
                }
            }
        }
    },
...
}